requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.0
django-cors-headers==3.2.0
django-crispy-forms==1.8.1
django-filter==2.0.0
django-formtools==2.1
django-guardian==1.4.9
django-import-export==2.0
django-reversion==3.0.2
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-extensions==0.4.0
et-xmlfile==1.0.1
freetype-py==2.1.0.post1
future==0.18.2
httplib2==0.9.2
idna==2.8
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10
Markdown==3.0.1
MarkupPy==1.14
MarkupSafe==1.1.0
MySQL-python==1.2.5
mysqlclient==1.4.6
numpy==1.17.4
odfpy==1.4.0
opencv-python==4.2.0.32
openpyxl==2.6.4
pandas==1.0.3
Pillow==6.2.1
pipenv==2018.11.26
Pygments==2.2.0
PyJWT==1.6.4
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.2
requests==2.22.0
scipy==1.1.0
six==1.13.0
sqlparse==0.3.0
tablib==0.14.0
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.25.7
virtualenv==16.7.8
virtualenv-clone==0.5.3
virtualenvwrapper-win==1.2.5
xadmin==0.6.1
xlrd==1.2.0
xlwt==1.3.0
